GPT-2-Simple_Model_Finetuning
This is the notebook from Max Woolf, where he uses the GPT-2-simple library to fine-tune the GPT-2. With GPT-2-simple it is possible to fine-tune the 355M-Model whereas with aitextgen apparently only the 124M-Version works. I tried the fine-tuning with the SQuAD-dataset and added only some minor changes to the notebook.